Refactoring folder name: exact -> general_state
